Move download content to own package
Create a download package for the repo, bundle, and
update content. This package contains the functionality
to download the desired content to the cache location.
